Dockerfile and script to start a local leap/eosio node
Adds a Dockerfile and a script to run a local leap (eosio) node
suitable for testing.

The script uses a token with 8 decimals and symbol `WAX` as a system
currency. This is because we want to start testing WAX and it's nice
to have everything close to that.

The Docker image installs the `leap` package which most importantly
contains `nodeos` and `cleos`. It also compiles the reference smart
contracts that are required to bootstrap a node.

The script follows the BIOS boot sequence. The steps are partially
taken from the old EOSIO documentation and the DUNE project belonging
to Antelope.

In this commit there is one small problem with the deployment of
`eosio.system` sometimes failing in the script. If that happens the
step for the deployment of `eosio.system` and the call of `eosio:init`
have to be done manually again.

All accounts are created with the same public/private key. We want to
use this for testing, so security isn't really a concern.