Used --replay-blockchain instead of --hard-replay
The latter one does not work on OSX for whatever reason (it complains
about not being able to rename a folder since it's not empty on
container startup). In manual tests `--replay-blockchain` worked and
did the same we want (clean the dirty database flag).

Also recommends another port as 7000 is used by default on OSX. Good
job, OSX!